+++ modules/shared/core/src/sap/commerce/toolset/psi/RangeAwareContent.kt
@@ -16,19 +16,11 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-/*
- * ----------------------------------------------------------------
- * --- WARNING: THIS FILE IS GENERATED AND WILL BE OVERWRITTEN! ---
- * ----------------------------------------------------------------
- */
-package sap.commerce.toolset.impex.psi;
+package sap.commerce.toolset.psi
 
-import java.util.List;
-import org.jetbrains.annotations.*;
-import com.intellij.psi.PsiElement;
+import com.intellij.openapi.util.TextRange
 
-public interface ImpExHeaderTypeName extends PsiElement {
-
-  @Nullable ImpExHeaderLine getHeaderLine();
-
-}
+data class RangeAwareContent(
+    val textRange: TextRange,
+    val content: String
+)